[01]
- Renomear namespace
- Modificar factory para receber um tipo e retornar seu concreto
