+++ DeadFishStudio.CoreLibrary.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.28010.2041
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GianLuca.Domain.Core", "GianLuca.Domain.Core\GianLuca.Domain.Core.csproj", "{969D8A15-7BC9-40FA-BFB6-02D9BABAF2E0}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DeadFishStudio.CoreLibrary", "DeadFishStudio.CoreLibrary\DeadFishStudio.CoreLibrary.csproj", "{969D8A15-7BC9-40FA-BFB6-02D9BABAF2E0}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GianLuca.Domain.Core.UnitTest", "GianLuca.Domain.Core.UnitTest\GianLuca.Domain.Core.UnitTest.csproj", "{21D325F8-9B29-42AE-BC37-124B976A30A5}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DeadFishStudio.CoreLibrary.UnitTest", "DeadFishStudio.CoreLibrary.UnitTest\DeadFishStudio.CoreLibrary.UnitTest.csproj", "{21D325F8-9B29-42AE-BC37-124B976A30A5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
