+++ DeadFishStudio.CoreLibrary/Interfaces/Repository/IBaseRepositoryAsync.cs
@@ -1,47 +1,47 @@
-﻿// <copyright file="IBaseService.cs" company="Gian Luca da Silva Figueiredo">
+﻿// <copyright file="IBaseRepositoryAsync.cs" company="Gian Luca da Silva Figueiredo">
 // Copyright (c) Gian Luca da Silva Figueiredo. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 // </copyright>
 
-namespace GianLuca.Domain.Core.Interfaces.Services
+namespace DeadFishStudio.CoreLibrary
 {
     using System;
     using System.Collections.Generic;
-    using GianLuca.Domain.Core.Entity;
+    using System.Threading.Tasks;
 
     /// <summary>
-    /// Interface síncrona de serviço.
+    /// Interface assíncrona para salvamento no banco de dados.
     /// </summary>
     /// <typeparam name="T">Entidade que será salva.</typeparam>
-    public interface IBaseService<T>
+    public interface IBaseRepositoryAsync<T>
         where T : BaseEntity
     {
         /// <summary>
-        /// Adiciona uma nova entidade no banco de dados.
+        /// Adiciona nova entidade no banco de dados de forma assíncrona.
         /// </summary>
         /// <param name="item">Entidade a ser salva.</param>
         /// <returns>Entidade salva.</returns>
-        T AddItem(T item);
+        Task<T> AddItemAsync(T item);
 
         /// <summary>
-        /// Retorna uma entidade com base em um identificador.
+        /// Retorna uma entidade com base em um identificador de forma assíncrona.
         /// </summary>
         /// <param name="id">Identificador da entidade.</param>
         /// <returns>Entidade encontrada.</returns>
-        T GetItem(Guid id);
+        Task<T> GetItemAsync(Guid id);
 
         /// <summary>
-        /// Retorna todas as entidades do banco de dados.
+        /// Retorna todas as entidades do banco de dados de forma assíncrona.
         /// </summary>
         /// <returns>Todas as entidades.</returns>
-        IEnumerable<T> GetAllItems();
+        Task<IEnumerable<T>> GetAllItemsAsync();
 
         /// <summary>
-        /// Atualiza um item com base em um identificador passado.
+        /// Atualiza uma entidade com base em um identificador passado.
         /// </summary>
         /// <param name="id">Identificador da entidade.</param>
         /// <param name="item">Entidade a ser atualizada.</param>
-        /// <returns>Entidade atualizada.</returns>
+        /// <returns>Entidade atualizado.</returns>
         T UpdateItem(Guid id, T item);
 
         /// <summary>
