+++ DeadFishStudio.CoreLibrary/Interfaces/Services/IBaseService.cs
@@ -1,23 +1,22 @@
-﻿// <copyright file="IBaseRepository.cs" company="Gian Luca da Silva Figueiredo">
+﻿// <copyright file="IBaseService.cs" company="Gian Luca da Silva Figueiredo">
 // Copyright (c) Gian Luca da Silva Figueiredo. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 // </copyright>
 
-namespace GianLuca.Domain.Core.Interfaces.Repository
+namespace DeadFishStudio.CoreLibrary
 {
     using System;
     using System.Collections.Generic;
-    using GianLuca.Domain.Core.Entity;
 
     /// <summary>
-    /// Interface síncrona para salvamento no banco de dados.
+    /// Interface síncrona de serviço.
     /// </summary>
     /// <typeparam name="T">Entidade que será salva.</typeparam>
-    public interface IBaseRepository<T>
+    public interface IBaseService<T>
         where T : BaseEntity
     {
         /// <summary>
-        /// Adiciona nova entidade no banco de dados.
+        /// Adiciona uma nova entidade no banco de dados.
         /// </summary>
         /// <param name="item">Entidade a ser salva.</param>
         /// <returns>Entidade salva.</returns>
@@ -37,11 +36,11 @@ namespace GianLuca.Domain.Core.Interfaces.Repository
         IEnumerable<T> GetAllItems();
 
         /// <summary>
-        /// Atualiza uma entidade com base em um identificador.
+        /// Atualiza um item com base em um identificador passado.
         /// </summary>
         /// <param name="id">Identificador da entidade.</param>
         /// <param name="item">Entidade a ser atualizada.</param>
-        /// <returns>Entidade atualizado.</returns>
+        /// <returns>Entidade atualizada.</returns>
         T UpdateItem(Guid id, T item);
 
         /// <summary>
