+++ DeadFishStudio.CoreLibrary/stylecop.json
@@ -13,7 +13,6 @@
       "useTabs": false
     },
     "spacingRules": {
-
     },
     "readabilityRules": {
       "allowBuiltInTypeAliases": false
